# apps/web/app/api/auth/complete/route.ts
import { NextRequest, NextResponse } from "next/server";
import { IdentityCompletionSchema, ReturnTargetSchema } from "@qelvora/api";
import {
  applicationOrigin,
  sameRequestOrigin,
} from "../../../../lib/request-origin";
import {
  continuationCookie,
  continuationReturnCookie,
  sessionCookie,
  cookieOptions,
  platformFetch,
} from "../../../../lib/session";

function retry(request: NextRequest, error: string) {
  const saved = ReturnTargetSchema.safeParse(
    request.cookies.get(continuationReturnCookie)?.value,
  );
  const target = new URL("/auth/continue", applicationOrigin(request));
  target.searchParams.set("returnTo", saved.success ? saved.data : "/home");
  target.searchParams.set("error", error);
  const response = NextResponse.redirect(target, 303);
  response.headers.set("Cache-Control", "no-store");
  response.cookies.delete(continuationCookie);
  return response;
}

export async function POST(request: NextRequest) {
  if (!sameRequestOrigin(request))
    return Response.json(
      { error: { message: "Start sign-in from this app." } },
      { status: 403 },
    );
  const form = await request.formData();
  const continuationId = request.cookies.get(continuationCookie)?.value;
  if (!continuationId || form.get("continuationId") !== continuationId)
    return retry(request, "continuation_expired");
  return complete(request, continuationId, form.get("code"), form.get("state"));
}

export async function GET(request: NextRequest) {
  const continuationId = request.cookies.get(continuationCookie)?.value;
  const query = request.nextUrl.searchParams;
  if (
    !continuationId ||
    query.getAll("continuationId").length !== 1 ||
    query.get("continuationId") !== continuationId ||
    query.getAll("code").length !== 1 ||
    query.getAll("state").length !== 1
  )
    return retry(request, "continuation_expired");
  return complete(
    request,
    continuationId,
    query.get("code"),
    query.get("state"),
  );
}

async function complete(
  request: NextRequest,
  continuationId: string,
  code: FormDataEntryValue | null,
  state: FormDataEntryValue | null,
) {
  try {
    const response = await platformFetch(
      "/v1/identity/complete",
      {
        method: "POST",
        headers: { "Content-Type": "application/json" },
        body: JSON.stringify({
          continuationId,
          code,
          ...(typeof state === "string" ? { state } : {}),
        }),
      },
      false,
    );
    if (!response.ok) {
      const failure = await response.json();
      return retry(
        request,
        failure.error?.code === "continuation_expired"
          ? "continuation_expired"
          : "continuation_failed",
      );
    }
    const result = IdentityCompletionSchema.parse(await response.json());
    const target = result.session.fan
      ? result.returnTo
      : `/onboarding/handle?returnTo=${encodeURIComponent(result.returnTo)}`;
    const redirect = NextResponse.redirect(
      new URL(target, applicationOrigin(request)),
      303,
    );
    redirect.cookies.set(sessionCookie, result.token, {
      ...cookieOptions,
      maxAge: 7 * 86400,
    });
    redirect.cookies.delete(continuationCookie);
    redirect.cookies.delete(continuationReturnCookie);
    return redirect;
  } catch {
    return retry(request, "continuation_failed");
  }
}